Collapse handler tests.
