+++ static/testdata/handlers/single/aggregate.go
