--- static/testdata/projections/single-projection-app/app.go
@@ -1,14 +1,0 @@
-package app
-
-import "github.com/dogmatiq/dogma"
-
-// App implements dogma.Application interface.
-type App struct{}
-
-// Configure configures the behavior of the engine as it relates to this
-// application.
-func (App) Configure(c dogma.ApplicationConfigurer) {
-	c.Identity("<single-projection-app>", "009eed19-bc2b-4a8c-8aa1-64b57da01e6e")
-
-	c.RegisterProjection(ProjectionHandler{})
-}
